Made the notifications a proper strategy pattern
I was having a look at this and realised there was too much messaging stuff in the implementation of the
notifications. The message is now created inside the notification class and that is sent to the individual
notifier. This is more correct implementation of the strategy pattern (I don't really know what the fuck
I was thinking when I wrote it originally)